+++ src/properties.rs
@@ -35,3 +35,4 @@ prop!(evaporation_heat, "evaporation_heat");
 prop!(electronic_configuration, "electronic_configuration");
 prop!(discoverers, "discoverers");
 prop!(discovery_location, "discovery_location");
+prop!(oxidation_states, "oxidation_states");